requirements.txt
streamlit
opencv-python
pillow
pytesseract
scikit-image
numpy
